Silence a warning on the Alpha.
Reviewed by:	md5(1)

+++ usr.bin/cap_mkdb/cap_mkdb.c
@@ -177,7 +177,8 @@ db_build(ifiles)
 			break;
 		case 2:
 			((char *)(data.data))[0] = TCERR;
-			warnx("record not tc expanded: %.*s", p - bp, bp);
+			warnx("record not tc expanded: %.*s", (int)(p - bp),
+			    bp);
 			break;
 		}
